Rename ProfileIO to V1ProfileConversion
Remove all write functions from V1ProfileConversion
Replace 'team config' with 'Zowe client config'
