+++ packages/imperative/src/profiles/src/utils/index.ts
@@ -9,5 +9,5 @@
 *
 */
 
-export * from "./ProfileIO";
+export * from "./V1ProfileConversion";
 export * from "./ProfileUtils";
